+++ src/main/java/com/sprint/mission/discodeit/event/listener/KafkaProduceRequiredEventListener.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.event.kafka;
+package com.sprint.mission.discodeit.event.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sprint.mission.discodeit.event.BinaryContentSaveFailEvent;
