+++ src/main/java/com/sprint/mission/discodeit/event/listener/NotificationRequiredTopicListener.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.event.kafka;
+package com.sprint.mission.discodeit.event.listener;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
